requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
dbfread
pandas
openpyxl
python-multipart
